requirements.txt
Flask==2.3.3 
langchain==0.2.16 
langchain-ollama==0.1.3 
pydantic==2.9.2 
python-dotenv==1.0.1 
flask-cors==5.0.0

